requirements.txt
Flask==3.0.0
gunicorn==21.2.0
Werkzeug==3.0.1
click>=8.1.3
itsdangerous>=2.1.2
Jinja2>=3.1.2
supabase==1.0.4
python-dotenv==1.0.0
Flask-Login==0.6.3
requests==2.31.0
oauthlib==3.2.2
pyOpenSSL==23.3.0
